pyproject.toml
[project]
name = "vanilla-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain-openai>=0.3.27",
    "python-dotenv>=1.1.1",
]
